moved advanced search code to AdvancedSearch.js
